requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.8.2
Brotli==1.0.7
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.0
cssselect==1.1.0
daphne==2.5.0
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
distlib==0.3.0
distro==1.4.0
Django==3.1
django-filter==2.3.0
django-modelcluster==5.0.2
django-plotly-dash==1.4.2
django-redis==4.12.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.1
dpd-components==0.1.0
draftjs-exporter==2.1.7
ebird-api==3.0.5
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
hiredis==1.1.0
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
isort==5.4.2
itemadapter==0.1.0
itemloaders==1.0.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
l18n==2018.5
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.19.1
openpyxl==3.0.5
packaging==20.3
pandas==1.1.1
parsel==1.6.0
pep517==0.8.2
Pillow==7.2.0
plotly==4.9.0
progress==1.5
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
queuelib==1.5.0
redis==3.5.3
requests==2.22.0
retrying==1.3.3
Scrapy==2.3.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
w3lib==1.22.0
wagtail==2.10
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.4
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
zope.interface==5.1.0
